streamlit/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
groq>=0.11.0
python-dotenv>=1.0.0